Visual "cy" mapping
In VISUAL mode "cy" grabs all the visually-selected text and pipes it to the
client's pbcopy command via ssh. If successful, it does a straight-forward
"y" operation on that same visual selection. No operator-pending mappings or
integration with custom registers.

The VISUAL cy command only succeeds if the g:VimPbcopyHost variable is
defined, else it prints out an error message directing user to set the
variable in their .vimrc file.