Support a g:vim_pbcopy_escape_backslashes flag
If the variable is set and is truthy, the plugin will always call
shellescape(escape("TEXT YOU WANT TO COPY")). E.g.

    let g:vim_pbcopy_escape_backslashes = 1

If the variable  is set and is falsey, the plugin will always call
shellescape("TEXT YOU WANT TO COPY"). E.g.,

    let g:vim_pbcopy_escape_backslashes = 0

If the variable is UNSET, the plugin will fall back to it's old and flawed
method of determining whether or not to escape back slashes. Note I am
seeing different behavior on different Mac OS X systems. It's hard to nail
down exactly what the root cause is, which is why I'm exposing a flag that
users can just set in their ~/.vimrc and not worry about.