Stub implementation for local, ssh, http protocols
Checks for a g:vim_pbcopy_protocol override to determine which protocol to
use, otherwise just uses "local" or "ssh" depending on whether we're running
locally or not. The override can be "local", "ssh", or "http".

No implementation for the http protocol yet, but the plan is to run a curl
command that sends the copied text in a JSON payload to an HTTP server.
Something like { "text": "the text you just copied" }. That HTTP server
would have to be running locally so that it can spawn the pbcopy command.
This will make it easier to copy text from a remote machine from which you
do NOT have reverse SSH access back into your local Mac machine. You could
SSH in to the remote machine with local port forwarding enabled and then
make the curl request through the SSH tunnel.